+++ packages/adhere-ui-searchtable/es/searchtable.d.ts
@@ -4,6 +4,7 @@ import { ColumnType, FilterValue, SorterResult, TableCurrentDataSource, TablePag
 import Suspense from '@baifendian/adhere-ui-suspense';
 import { ISearchTableProps, ISearchTableState } from './types';
 import SearchForm from './searchform';
+export declare const selectorPrefix = "adhere-ui-searchtable";
 /**
  * SearchTable
  * @class SearchTable
@@ -15,7 +16,11 @@ declare abstract class SearchTable extends Suspense<ISearchTableProps, ISearchTa
     static SearchForm: SearchForm;
     static NUMBER_GENERATOR_RULE_ALONE: symbol;
     static NUMBER_GENERATOR_RULE_CONTINUITY: symbol;
+    static ROW_SELECTION_NORMAL_MODE: symbol;
+    static ROW_SELECTION_CONTINUOUS_MODE: symbol;
     protected tableWrapRef: RefObject<HTMLDivElement>;
+    private components;
+    private columnObserver;
     /**
      * isShowNumber - 表格是否显示序号
      * @return boolean
@@ -25,11 +30,15 @@ declare abstract class SearchTable extends Suspense<ISearchTableProps, ISearchTa
      * getTableNumberColumnWidth - 表格序号列的宽度
      * @return number
      */
-    abstract getTableNumberColumnWidth(): number;
+    abstract getTableNumberColumnWidth(): Symbol;
     /**
      * getNumberGeneratorRule - 获取符号列的生成规则
      */
-    abstract getNumberGeneratorRule(): string;
+    abstract getNumberGeneratorRule(): Symbol;
+    /**
+     * getRowSelectionMode - 获取全选的生模式
+     */
+    abstract getRowSelectionMode(): string;
     /**
      * getRowKey - 获取表格的主键属性
      * @return string
@@ -104,6 +113,7 @@ declare abstract class SearchTable extends Suspense<ISearchTableProps, ISearchTa
      * @protected
      */
     protected renderTableNumberColumn(number: string | undefined, params: {
+        value: any;
         record: object;
         index: number;
     }): JSX.Element;
@@ -140,6 +150,11 @@ declare abstract class SearchTable extends Suspense<ISearchTableProps, ISearchTa
      * @return Array<any>
      */
     protected getTableColumns(): Array<any>;
+    /**
+     * renderColumnSetting
+     * @description 创建列设置组件
+     */
+    renderColumnSetting(): React.ReactElement;
     /**
      * renderSearchFooter - 渲染查询工具栏
      * @return React.ReactElement