+++ packages/adhere-ui-searchtable/lib/Extension/ColumnResizable/SearchTableResizableColumnItem.d.ts
@@ -1,0 +1,2 @@
+declare const _default: (context: any, column: any) => any;
+export default _default;