+++ packages/adhere-website/src/components/ui/permission/index.jsx
@@ -41,7 +41,7 @@ export default () => {
     <Button
       type="primary"
       onClick={() => {
-        const el = MessageDialog.Modal({
+        const {close} = MessageDialog.Modal({
           config: {
             title: '权限所有设置',
             width: 200,
@@ -54,7 +54,8 @@ export default () => {
                     setPermission(val.split(','));
                     setAllPermission(getPermission());
                   }
-                  MessageDialog.close(el);
+                  // MessageDialog.close(el);
+                  close();
                 }}
               >
                 确定
@@ -74,7 +75,7 @@ export default () => {
 
     <Button
       onClick={() => {
-        const el = MessageDialog.Modal({
+        const {close} = MessageDialog.Modal({
           config: {
             title: '权限当前设置',
             width: 200,
@@ -86,7 +87,8 @@ export default () => {
                   if (val) {
                     setCurPermission(val.split(','));
                   }
-                  MessageDialog.close(el);
+                  // MessageDialog.close(el);
+                  close();
                 }}
               >
                 确定
@@ -117,7 +119,7 @@ export default () => {
               <Button
                 type="primary"
                 onClick={() => {
-                  const el = MessageDialog.Modal({
+                  const { close } = MessageDialog.Modal({
                     config: {
                       title: '权限所有设置',
                       width: 200,
@@ -130,7 +132,8 @@ export default () => {
                               setPermission(val.split(','));
                               setAllPermission(getPermission());
                             }
-                            MessageDialog.close(el);
+                            // MessageDialog.close(el);
+                            close();
                           }}
                         >
                           确定
@@ -150,7 +153,7 @@ export default () => {
 
               <Button
                 onClick={() => {
-                  const el = MessageDialog.Modal({
+                  const { close } = MessageDialog.Modal({
                     config: {
                       title: '权限当前设置',
                       width: 200,
@@ -162,7 +165,8 @@ export default () => {
                             if (val) {
                               setCurPermission(val.split(','));
                             }
-                            MessageDialog.close(el);
+                            // MessageDialog.close(el);
+                            close();
                           }}
                         >
                           确定