+++ packages/adhere-website/src/components/ui/polygonselection/imageselect.jsx
@@ -201,7 +201,8 @@ export default () => {
                                 localData.splice(index, 1);
                                 Preferences.putObjectByLocal('polygonSelectionData', localData);
                                 message.success('删除成功！');
-                                MessageDialog.close(el);
+                                // MessageDialog.close(el);
+                                close();
                               }
                             }}
                           >
@@ -212,7 +213,8 @@ export default () => {
                             href="#"
                             onClick={() => {
                               message.success('操作成功！');
-                              MessageDialog.close(el);
+                              // MessageDialog.close(el);
+                              close();
                               const { data } = record;
                               polygonSelection.current.setHistoryData(data);
                               polygonSelection.current.clearDraw();
@@ -227,7 +229,7 @@ export default () => {
                   },
                 ];
 
-                const el = MessageDialog.Modal({
+                const { close } = MessageDialog.Modal({
                   config: {
                     title: '设置数据',
                   },
@@ -358,7 +360,7 @@ export default () => {
           </li>
         </ul>
       </div>
-      <div className={styles.Auto} ref={autoRef}></div>
+      <div className={styles.Auto} ref={autoRef} />
     </div>
   );
 };