+++ packages/adhere-website/src/components/ui/searchtable/columnSettingTable.jsx
@@ -1,0 +1,22 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import React from 'react';
+
+import Table from './table';
+import styles from './fixedTableSpaceBetweenTable.less';
+
+/**
+ * ColumnSettingTable
+ * @classdesc
+ */
+class ColumnSettingTable extends Table {
+  renderTableHeader() {
+    return (
+      <div className={styles.Header}>
+        <h3>查询表格</h3>
+        <div>{this.renderColumnSetting()}</div>
+      </div>
+    );
+  }
+}
+
+export default ColumnSettingTable;